requirements.txt
torch~=2.5.0
flask~=3.0.3
diffusers~=0.31.0
pillow~=11.0.0
numpy~=2.0.2
pip~=24.2
wheel~=0.44.0
filelock~=3.16.1
requests~=2.32.3
colorama~=0.4.6
PyYAML~=6.0.2
zipp~=3.20.2
MarkupSafe~=3.0.1
werkzeug~=3.0.4
click~=8.1.7
jinja2~=3.1.4
blinker~=1.8.2
itsdangerous~=2.2.0
setuptools~=75.1.0
regex~=2024.9.11
sympy~=1.13.1
mpmath~=1.3.0
tqdm~=4.66.5
psutil~=6.1.0
packaging~=24.1
urllib3~=2.2.3
fsspec~=2024.10.0
certifi~=2024.8.30
networkx~=3.2.1
idna~=3.10
accelerate~=1.0.1
transformers~=4.45.2
safetensors~=0.4.5
tomli~=2.0.1
autocommand~=2.2.2
inflect~=7.3.1
typeguard~=4.3.0
platformdirs~=4.2.2
tokenizers~=0.20.1